requirements.txt
rich
requests
dnspython
python-whois
pyfiglet
psutil
